requirements.txt
h5py==2.10.0
numpy==1.18.5
# python,tensorflow和keras的版本对应关系
# https://docs.floydhub.com/guides/environments/
tensorflow-gpu==2.2.0
keras==2.3.1
opencv-python==4.5.1.48
matplotlib==3.3.4
Pillow==8.1.0
pymysql==1.0.2
